LayoutTests/fast/innerHTML/innerHTML-uri-resolution.html

<!DOCTYPE html>
<html>
<head>
<link rel="stylesheet" href="../js/resources/js-test-style.css">
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div id="container"></div>
<script>
  description('Checks that URIs used in styles that are set via innerHTML are resolved against the current document\'s base URI.');
  
  document.getElementById('container').innerHTML = 
      '<span id="test-span" style="background-image: url(image.png)"></span>';
  
  var currentUri = document.location.href;
  var currentPath = currentUri.substring(0, currentUri.lastIndexOf('/'));
  var cssUriWasResolvedAgainstDocumentUri = document.getElementById("test-span").style.backgroundImage == 'url(' + currentPath + '/image.png)';
  
  // Can't log the actual path since it's different depending on where the test is run.
  shouldBeTrue('cssUriWasResolvedAgainstDocumentUri');

  var successfullyParsed = true;
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
